+++ src/commands/randomSubWeapon.ts
